+++ packages/backend/src/storage/storage.ts
@@ -35,8 +35,9 @@ import { StorageEvents } from './types'
 import { IpfsFilesManager, IpfsFilesManagerEvents } from './ipfsFileManager'
 
 import { CID } from 'multiformats/cid'
-import { ChannelMessage, ConnectionProcessInfo, FileMetadata, NoCryptoEngineError, PublicChannel, PushNotificationPayload, SaveCertificatePayload, SocketActionTypes, User } from '@quiet/types'
+import { ChannelMessage, ConnectionProcessInfo, DeleteFilesFromChannelSocketPayload, FileMetadata, NoCryptoEngineError, PublicChannel, PushNotificationPayload, SaveCertificatePayload, SocketActionTypes, User } from '@quiet/types'
 import { isDefined } from '@quiet/common'
+import fs from 'fs'
 
 const log = logger('db')
 
@@ -44,7 +45,6 @@ const { createPaths, removeDirs, removeFiles, getUsersAddresses } = await import
 export class Storage extends EventEmitter {
   public quietDir: string
   public peerId: PeerId
-  public ownerPeerId: PeerId
   protected ipfs: IPFS
   protected orbitdb: OrbitDB
   public channels: KeyValueStore<PublicChannel>
@@ -554,13 +554,14 @@ export class Storage extends EventEmitter {
     return db
   }
 
-  public async deleteChannel(payload: { channelId: string }) {
+  public async deleteChannel(payload: { channelId: string; ownerPeerId: string }) {
     console.log('deleting channel storage', payload)
-    const { channelId } = payload
+    const { channelId, ownerPeerId } = payload
     // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
     await this.channels.load({ fetchEntryTimeout: 15000 })
     const channel = this.channels.get(channelId)
-    if (channel) {
+    const isOwner = ownerPeerId === this.peerId.toString()
+    if (channel && isOwner) {
       await this.channels.del(channelId)
     }
     let repo = this.publicChannelsRepos.get(channelId)
@@ -590,7 +591,8 @@ export class Storage extends EventEmitter {
     // await this.deleteChannelFiles(files)
     // await this.deleteChannelMessages(hashes)
     this.publicChannelsRepos.delete(channelId)
-    this.emit(StorageEvents.CHANNEL_DELETION_RESPONSE, payload)
+    const responsePayload = { channelId: payload.channelId }
+    this.emit(StorageEvents.CHANNEL_DELETION_RESPONSE, responsePayload)
   }
 
   public async deleteChannelFiles(files: FileMetadata[]) {
@@ -855,5 +857,47 @@ export class Storage extends EventEmitter {
     }
 
     return this.userNamesMap.get(publicKey)
+  }
+
+  public async deleteFilesFromChannel(payload: DeleteFilesFromChannelSocketPayload) {
+    const { messages } = payload
+    Object.keys(messages).map(async(key) => {
+      const message = messages[key]
+      if (message?.media?.path) {
+        const mediaPath = message.media.path
+        log('deleteFilesFromChannel : mediaPath', mediaPath)
+        const isFileExist = await this.checkIfFileExist(mediaPath)
+        log(`deleteFilesFromChannel : isFileExist- ${isFileExist}`)
+        if (isFileExist) {
+            fs.unlink(mediaPath, unlinkError => {
+              if (unlinkError) {
+                log(`deleteFilesFromChannel : unlink error - ${unlinkError}`)
+              }
+            })
+        } else {
+          log(`deleteFilesFromChannel : file dont exist - ${mediaPath}`)
+        }
+      }
+    })
+  }
+
+  public async checkIfFileExist(filepath: string): Promise<boolean> {
+      return await new Promise((resolve) => {
+        fs.access(filepath, fs.constants.F_OK, error => {
+          resolve(!error)
+        })
+      })
+  }
+
+  private async deleteFilesFromTemporaryDir() {
+    const temporaryFilesDirectory = path.join(this.quietDir, '/../', 'temporaryFiles')
+    fs.readdir(temporaryFilesDirectory, (readDirErr, files) => {
+      if (readDirErr) log(`deleteFilesFromTemporaryDir : readdir error - ${readDirErr}`)
+      for (const file of files) {
+        fs.unlink(path.join(temporaryFilesDirectory, file), unlinkError => {
+          if (unlinkError) log(`deleteFilesFromTemporaryDir : unlink error - ${unlinkError}`)
+        })
+      }
+    })
   }
 }